# auth-service/swagger.yaml
openapi: 3.0.3
info:
  title: Auth Service API
  description: |
    Handles user registration, login, JWT token issuance and validation.
    All other microservices call `GET /auth/validate-token` to verify Bearer tokens.
  version: 1.0.0
  contact:
    name: QuickBite Team

servers:
  - url: http://localhost:3001
    description: Local development
  - url: http://localhost/api/auth
    description: Via Nginx API Gateway (local)

tags:
  - name: Auth
    description: Authentication & authorisation

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string

    UserPublic:
      type: object
      properties:
        id:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d1"
        name:
          type: string
          example: "Jane Doe"
        email:
          type: string
          format: email
          example: "jane@example.com"
        role:
          type: string
          enum: [customer, restaurant_owner, delivery_driver, admin]
          example: customer

    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
          example: "Jane Doe"
        email:
          type: string
          format: email
          example: "jane@example.com"
        password:
          type: string
          minLength: 6
          example: "secret123"
        role:
          type: string
          enum: [customer, restaurant_owner, delivery_driver]
          default: customer

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "jane@example.com"
        password:
          type: string
          example: "secret123"

    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                token:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                user:
                  $ref: '#/components/schemas/UserPublic'

paths:
  /health:
    get:
      tags: [Auth]
      summary: Health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive a JWT
      description: Rate limited to 10 requests per 15 minutes per IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many login attempts

  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout (client should discard the token)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  /auth/validate-token:
    get:
      tags: [Auth]
      summary: Validate a JWT — used by other microservices
      description: Called internally by restaurant, order, and delivery services to verify tokens.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            type: object
                            properties:
                              userId:
                                type: string
                              email:
                                type: string
                              role:
                                type: string
        '401':
          description: Invalid or expired token

  /auth/profile:
    get:
      tags: [Auth]
      summary: Get the authenticated user's profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile returned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserPublic'
        '401':
          description: Unauthorized

    put:
      tags: [Auth]
      summary: Update name or password
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                password:
                  type: string
                  minLength: 6
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized